refactor(api): introduce a casting utility class for casting WMI result objects
This change introduces a dedicated utility for safely converting WMI query
results to specific types.

 The `installedSize` field in `ProcessorCache` was also updated to `Long` from `Integer` to match Microsoft's documentation.

+++ src/main/java/org/pinetree/entity/processor/ProcessorCache.java
@@ -49,7 +49,7 @@ public class ProcessorCache {
 
     @SerializedName("InstalledSize")
     @Nullable
-    Integer installedSize;
+    Long installedSize;
 
     @SerializedName("Associativity")
     @Nullable
